+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes for Opayo for Craft Commerce
 
+## 1.0.2 - 2022-06-10
+### Added
+- Craft 4 Compatibility
+- Commerce 4 Compatibility
+
+## 1.0.1 - 2022-03-28
+### Changed
+- Logo updated.
+
 ## 1.0.0 - 2022-03-07
 
 - Initial release.